move some util tests from comments into separate files
