+++ util/mt19937ar.c
@@ -230,24 +230,3 @@ double nlopt_nrand(double mean, double stddev)
     return mean + v1 * sqrt(-2 * log(s) / s) * stddev;
   }
 }
-
-#if 0
-#include <stdio.h>
-int main(void)
-{
-    int i;
-    uint32_t init[4]={0x123, 0x234, 0x345, 0x456}, length=4;
-    init_by_array(init, length);
-    printf("1000 outputs of nlopt_genrand_int32()\n");
-    for (i=0; i<1000; i++) {
-      printf("%10lu ", nlopt_genrand_int32());
-      if (i%5==4) printf("\n");
-    }
-    printf("\n1000 outputs of genrand_real2()\n");
-    for (i=0; i<1000; i++) {
-      printf("%10.8f ", genrand_real2());
-      if (i%5==4) printf("\n");
-    }
-    return 0;
-}
-#endif
